+++ com.unity.multiplayer.mlapi/Runtime/Timing/ClientNetworkTimeProvider.cs
@@ -66,7 +66,7 @@ namespace MLAPI.Timing
         }
 
         /// <inheritdoc/>
-        public bool AdvanceTime(ref NetworkTime predictedTime, ref NetworkTime serverTime, float deltaTime)
+        public bool AdvanceTime(ref NetworkTime predictedTime, ref NetworkTime serverTime, double deltaTime)
         {
             // advance time
             predictedTime += deltaTime * PredictedTimeScale;
@@ -74,9 +74,9 @@ namespace MLAPI.Timing
 
             // time scale adjustment based on whether we are behind / ahead of the server in terms of inputs
             // This implementation uses RTT to calculate that without server input which is not ideal. In the future we might want to add a field to the protocol which allows the server to send the exact input buffers size to the client.
-            float lastReceivedSnapshotTime = m_NetworkStats.LastReceivedSnapshotTick.Time;
-            float targetServerTime = lastReceivedSnapshotTime - TargetClientBufferTime;
-            float targetPredictedTime = lastReceivedSnapshotTime + m_NetworkStats.Rtt + TargetServerBufferTime;
+            double lastReceivedSnapshotTime = m_NetworkStats.LastReceivedSnapshotTick.Time;
+            double targetServerTime = lastReceivedSnapshotTime - TargetClientBufferTime;
+            double targetPredictedTime = lastReceivedSnapshotTime + m_NetworkStats.Rtt + TargetServerBufferTime;
 
             // Reset timescale. Will be recalculated based on new values.
             bool reset = false;
